requirements.txt
PyGithub
python-dotenv
google-genai
requests